refactor(nucleus-tool-proxy): extract pod management handlers to pod_mgmt.rs
Move 342 lines of pod management, vsock, and runtime helpers from
main.rs (4785→4443 lines) to a dedicated pod_mgmt.rs module.

Add line ratchet for nucleus-tool-proxy/src/main.rs:
ceiling 4443, target 2500, step 1 per merge.

No behavioral changes — pure extraction with module-qualified call sites.
